App care Test
Using fragments, design a page with two buttons Save and View list. Where users can enter the name, department, and select a profile picture from gallery. Save all the data in the database using ROOM Database.

Design a page to show a list of details above mentioned (image with rounded shape) on the left and a delete icon on the right side. When clicked on Delete icon it should be removed from the list and database.